desktop_requirements.txt
pywebview==5.0.7
